[BUILD] Added check for spaces in absolute path to ARSDKBuildUtils
The ARSDK build system uses the absolute path to its directory in
various places, sometimes without proper escaping of the name. This
can cause issues when exec-ing other programs if the path has a space
in it, as it will be assumed to be two (or more) separate arguments.

To avoid this, CheckEnv.py now returns an error when the absolute path
to the ARSDKBuildUtils folder contains at least one space.

Change-Id: Ief007cd23c31be7fbba8bfe7e1bf24418a1538ab